Fix several big-endian bugs

+++ runtimes/native/src/framebuffer.h
@@ -6,7 +6,7 @@
 #define WIDTH 160
 #define HEIGHT 160
 
-void w4_framebufferInit (const uint16_t* drawColors, uint8_t* framebuffer);
+void w4_framebufferInit (const uint8_t* drawColors, uint8_t* framebuffer);
 
 void w4_framebufferClear ();
